llm_service/pyproject.toml
[project]
name = "llm_service"
version = "2.0.0"
dependencies = [
    "fastapi==0.124.4",
    "uvicorn==0.38.0",
    "pydantic==2.12.5",
    "python-dotenv==1.2.1",
    "groq",
]

[project.optional-dependencies]
gemini = ["google-generativeai>=0.8.6"]
openai = ["openai>=1.0.0"]
all = ["google-generativeai>=0.8.6", "openai>=1.0.0"]
